views and forms work now
finally the views and forms transition appropriately, save the fact that any time you try to submit a comment / post it attempts to direct to a template that doesnt exist. still troubleshooting that. work remaining is improving the interface and adding some scripting (i.e. pressing a button to raise post score) as well as general css implementation.

the users/accounts model needs to be done still and i will probably work on that next